написал тест проверки метода getFoodInformation
